Enforce integer units, add field validation with user-visible errors, and tighten input handling across recipe and ingredient forms
Change unitsPerRecipe from float to int across Recipe and FileHandler
so 'units' quantities can no longer be saved as fractional values. In
AddIngredientDialog and IngredientFrame, branch validation on the
ingredient's unit type so 'units' fields require whole numbers while
weight/volume fields accept decimals. In RecipeFrame, replace silent
getFloatFromString calls with new validateFloat and validateInt helpers
that surface a clear popup naming the offending field instead of
silently coercing bad input to zero.

+++ src/filep/FileHandler.java
@@ -208,7 +208,7 @@ public class FileHandler {
 						temp.setManpowerRate(Float.parseFloat(parts[6]));
 						temp.setManpowerMinutes(Float.parseFloat(parts[7]));
 						temp.setPackagingCostPerUnit(Float.parseFloat(parts[8]));
-						temp.setUnitsPerRecipe(Float.parseFloat(parts[9]));
+						temp.setUnitsPerRecipe(Integer.parseInt(parts[9]));
 						temp.setAllergens(unescapeNote(parts[10]));
 						temp.setPersonalNote(unescapeNote(parts[11]));
 						ingredientStart = 12;
